requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
pyjwt[crypto]==2.8.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
click==8.1.7
typer==0.9.0
rich==13.7.0
requests==2.31.0
aiohttp==3.9.1
aiosqlite==0.19.0
openai>=1.0.0
anthropic>=0.25.0
tiktoken>=0.5.0